# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "## A Full-Featured EDTECH API built with FIBER \u0026 ENT ORM.\n\n\u003c!-- ### WEBSOCKETS:\n\n#### Notifications\n\n- URL: `wss://{host}/api/v4/ws/notifications`\n\n- Requires authorization, so pass in the Bearer Authorization header.\n\n- You can only read and not send notification messages into this socket. --\u003e\n\n\n\u003c!-- #### Chats\n\n- URL: `wss://{host}/api/v4/ws/chats/{id}`\n- Requires authorization, so pass in the Bearer Authorization header.\n- Use chat_id as the ID for an existing chat or username if it's the first message in a DM.\n- You cannot read realtime messages from a username that doesn't belong to the authorized user, but you can surely send messages.\n- Only send a message to the socket endpoint after the message has been created or updated, and files have been uploaded.\n- Fields when sending a message through the socket:\n\n  ```json\n  { \"status\": \"CREATED\", \"id\": \"fe4e0235-80fc-4c94-b15e-3da63226f8ab\" }\n  ``` --\u003e",
        "title": "EDNET API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/profiles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint allows a user to view his/her profile`",
                "tags": [
                    "Profiles"
                ],
                "summary": "Get Your Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profiles.ProfileResponseSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/base.UnauthorizedErrorExample"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint allows a user to update his/her profile`",
                "tags": [
                    "Profiles"
                ],
                "summary": "Update Your Profile",
                "parameters": [
                    {
                        "maxLength": 300,
                        "minLength": 10,
                        "type": "string",
                        "example": "I'm the boss",
                        "name": "bio",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "2000-09-12",
                        "name": "dob",
                        "in": "formData"
                    },
                    {
                        "maxLength": 150,
                        "minLength": 10,
                        "type": "string",
                        "example": "John Doe",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string",
                        "example": "john-doe",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Profile picture to upload",
                        "name": "avatar",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profiles.ProfileResponseSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/base.UnauthorizedErrorExample"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/base.ValidationErrorExample"
                        }
                    }
                }
            }
        },
        "/profiles/courses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint allows a user to view his/her enrolled courses`",
                "tags": [
                    "Profiles"
                ],
                "summary": "Get Your Enrolled Courses",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Page Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter By Title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter By Status (inactive, active, completed, dropped)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter By Instructor's Name Or Username",
                        "name": "instructor",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter By Free Status",
                        "name": "isFree",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.CoursesResponseSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/base.UnauthorizedErrorExample"
                        }
                    }
                }
            }
        },
        "/profiles/courses/{slug}/progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint allows a user to get his/her course progress`",
                "tags": [
                    "Profiles"
                ],
                "summary": "Get Course Progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profiles.CourseProgressResponseSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/base.UnauthorizedErrorExample"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            }
        },
        "/profiles/leaderboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint retrieves the top 100 students by quiz score.`",
                "tags": [
                    "Profiles"
                ],
                "summary": "Get Leaderboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profiles.LeaderboardResponseSchema"
                        }
                    }
                }
            }
        },
        "/profiles/lessons/{slug}/progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint allows a user to get his/her lesson progress`",
                "tags": [
                    "Profiles"
                ],
                "summary": "Get Lesson Progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profiles.LessonProgressResponseSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/base.UnauthorizedErrorExample"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint allows a user to create or update a lesson progress`",
                "tags": [
                    "Profiles"
                ],
                "summary": "Create/Update Lesson Progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lesson Progress object",
                        "name": "lesson_progress",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profiles.LessonProgressInputSchema"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/profiles.LessonProgressResponseSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/base.UnauthorizedErrorExample"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/base.ValidationErrorExample"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "base.FieldData": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "This field is required"
                }
            }
        },
...
        "base.NotFoundErrorExample": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "The item was not found"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
...
        "base.UnauthorizedErrorExample": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Unauthorized user/Invalid credentials/Invalid Token"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
        "base.UserDataSchema": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://img.url"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "username": {
                    "type": "string",
                    "example": "john-doe"
                }
            }
        },
        "base.ValidationErrorExample": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/base.FieldData"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid Entry"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
        "config.PaginationResponse-courses_CourseListSchema": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/courses.CourseListSchema"
                    }
                },
                "items_count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "course.Difficulty": {
            "type": "string",
            "enum": [
                "beginner",
                "beginner",
                "intermediate",
                "advanced"
            ],
            "x-enum-varnames": [
                "DefaultDifficulty",
                "DifficultyBeginner",
                "DifficultyIntermediate",
                "DifficultyAdvanced"
            ]
        },
...
        "courses.CategoryOrTagSchema": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "courses.CourseListSchema": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/courses.CategoryOrTagSchema"
                },
                "created_at": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "difficulty": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/course.Difficulty"
                        }
                    ],
                    "example": "Beginner"
                },
                "discount_price": {
                    "type": "number"
                },
                "instructor": {
                    "$ref": "#/definitions/base.UserDataSchema"
                },
                "is_free": {
                    "type": "boolean",
                    "example": false
                },
                "is_published": {
                    "type": "boolean",
                    "example": false
                },
                "language": {
                    "type": "string",
                    "example": "English"
                },
                "lessons_count": {
                    "type": "integer",
                    "example": 20
                },
                "price": {
                    "type": "number",
                    "example": 19.99
                },
                "rating": {
                    "type": "number",
                    "example": 4.8
                },
                "slug": {
                    "type": "string",
                    "example": "go-programming-for-beginners"
                },
                "students_count": {
                    "type": "integer",
                    "example": 1200
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://ednet-images.com/courses/go.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Go Programming for Beginners"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "courses.CoursesResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/config.PaginationResponse-courses_CourseListSchema"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "profiles.CourseProgressResponseData": {
            "type": "object",
            "properties": {
                "percentage": {
                    "type": "number"
                }
            }
        },
        "profiles.CourseProgressResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/profiles.CourseProgressResponseData"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "profiles.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "total_score": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "profiles.LeaderboardResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/profiles.LeaderboardEntry"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "profiles.LessonProgressInputSchema": {
            "type": "object",
            "properties": {
                "is_completed": {
                    "type": "boolean"
                }
            }
        },
        "profiles.LessonProgressResponseData": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "profiles.LessonProgressResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/profiles.LessonProgressResponseData"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "profiles.ProfileResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/profiles.ProfileSchema"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "profiles.ProfileSchema": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "https://ednet-images.com/users/john-doe"
                },
                "bio": {
                    "type": "string",
                    "example": "I'm the boss"
                },
                "dob": {
                    "type": "string",
                    "example": "2000-09-12"
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/user.Role"
                        }
                    ],
                    "example": "student"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
        "user.Role": {
            "type": "string",
            "enum": [
                "student",
                "student",
                "instructor",
                "admin"
            ],
            "x-enum-varnames": [
                "DefaultRole",
                "RoleStudent",
                "RoleInstructor",
                "RoleAdmin"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer jwt_string' to correctly set the API Key",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}